backend/requirements.txt
# 热点新闻收集与推送工具依赖项

# 基础库
requests>=2.28.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
python-dateutil>=2.8.2
feedparser>=6.0.0

# 网页内容提取
newspaper3k>=0.2.8
trafilatura>=2.0.0
cloudscraper>=1.2.71

# 异步和并发
asyncio>=3.4.3

# LangChain相关
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-openai>=0.0.5

# 日志和工具
logging
pathlib>=1.0.1

# 数据处理

pickle-mixin>=1.0.2